fixed some warnings. may cause other warnings

+++ philosopher-king/Texture.cpp
@@ -59,7 +59,7 @@ unsigned char* Texture::loadPPM(const char* filename, int& width, int& height)
 
     // Read image data:
     rawData = new unsigned char[width * height * 3];
-    read = fread(rawData, width * height * 3, 1, fp);
+    read = (unsigned int)fread(rawData, width * height * 3, 1, fp);
     fclose(fp);
     if (read != 1)
     {
